feature get group messages & clear code.
