+++ client/offline_file.go
@@ -7,6 +7,10 @@ import (
 	"google.golang.org/protobuf/proto"
 )
 
+func init() {
+	decoders["OfflineFilleHandleSvr.pb_ftn_CMD_REQ_APPLY_DOWNLOAD-1200"] = decodeOfflineFileDownloadResponse
+}
+
 func (c *QQClient) buildOfflineFileDownloadRequestPacket(uuid []byte) (uint16, []byte) {
 	seq := c.nextSeq()
 	req := &cmd0x346.C346ReqBody{
